+++ test_data/generator.py
@@ -13,6 +13,26 @@ from legacy_password_hashing.password_hashing import (
     hash_ycc_password,
     verify_ycc_password,
 )
+from test_data.generator_config import (
+    BOATS_EXPORTED_JSON_FILE,
+    BOATS_JSON_FILE,
+    CURRENT_YEAR,
+    ENTRANCE_FEE_RECORDS_JSON_FILE,
+    FEE_RECORDS_JSON_FILE,
+    HELPER_TASK_CATEGORIES_JSON_FILE,
+    HELPER_TASK_HELPERS_JSON_FILE,
+    HELPER_TASKS_JSON_FILE,
+    LICENCES_JSON_FILE,
+    MEMBER_COUNT,
+    MEMBERS_JSON_FILE,
+    USERS_JSON_FILE,
+)
+from test_data.utils.helpers import (
+    generate_helper_task_categories,
+    generate_helper_task_helpers,
+    generate_helper_tasks,
+)
+from test_data.utils.licences import generate_licences
 from ycc_hull.db.entities import (
     BaseEntity,
     BoatEntity,
@@ -22,19 +42,6 @@ from ycc_hull.db.entities import (
     MemberEntity,
     UserEntity,
 )
-
-from test_data.generator_config import (
-    BOATS_EXPORTED_JSON_FILE,
-    BOATS_JSON_FILE,
-    CURRENT_YEAR,
-    ENTRANCE_FEE_RECORDS_JSON_FILE,
-    FEE_RECORDS_JSON_FILE,
-    LICENCES_JSON_FILE,
-    MEMBER_COUNT,
-    MEMBERS_JSON_FILE,
-    USERS_JSON_FILE,
-)
-from test_data.utils.licences import generate_licences
 
 MemberInfo = NamedTuple(
     "MemberInfo",
@@ -340,6 +347,34 @@ def _did_not_pay_fee_current_year(member_info: MemberInfo) -> bool:
     return not _paid_fee_current_year(member_info)
 
 
+def get_members_with_licence(
+    member_infos: List[MemberInfo], licence_info_id: int
+) -> Iterator[MemberInfo]:
+    return filter(
+        lambda member_info: _has_licence(member_info, licence_info_id), member_infos
+    )
+
+
+def get_members_without_licence(
+    member_infos: List[MemberInfo], licence_info_id: int
+) -> Iterator[MemberInfo]:
+    return filter(
+        lambda member_info: _lacks_licence(member_info, licence_info_id), member_infos
+    )
+
+
+def _has_licence(member_info: MemberInfo, licence_info_id: int) -> bool:
+    search = filter(
+        lambda licence: licence.licence_id == licence_info_id,
+        member_info.licences,
+    )
+    return any(search)
+
+
+def _lacks_licence(member_info: MemberInfo, licence_info_id: int) -> bool:
+    return not _has_licence(member_info, licence_info_id)
+
+
 def generate(force_regenerate: bool = False) -> None:
     if path.exists(MEMBERS_JSON_FILE) and not force_regenerate:
         print("== Skipping members")
@@ -387,12 +422,30 @@ def generate(force_regenerate: bool = False) -> None:
                 get_members_without_payment_current_year(member_infos[100:])
             ).user.logon_id
         )
+        print(
+            "Has M key: "
+            + next(get_members_with_licence(member_infos[100:], 9)).user.logon_id
+        )
+        print(
+            "Lacks M key: "
+            + next(get_members_without_licence(member_infos[100:], 9)).user.logon_id
+        )
 
     if path.exists(BOATS_JSON_FILE) and not force_regenerate:
         print("== Skipping boats")
     else:
         print("== Generating boats...")
         write_json_file(BOATS_JSON_FILE, generate_boats())
+
+    if path.exists(HELPER_TASKS_JSON_FILE) and not force_regenerate:
+        print("== Skipping helper data")
+    else:
+        print("== Generating helpers data (categories, tasks, helpers)...")
+        write_json_file(
+            HELPER_TASK_CATEGORIES_JSON_FILE, generate_helper_task_categories()
+        )
+        write_json_file(HELPER_TASKS_JSON_FILE, generate_helper_tasks())
+        write_json_file(HELPER_TASK_HELPERS_JSON_FILE, generate_helper_task_helpers())
 
 
 def regenerate() -> None: