+++ ycc_hull/models/dtos.py
@@ -1,5 +1,5 @@
 """
-API DTO Classes.
+General API DTO classes.
 """
 from datetime import date
 from typing import Optional
@@ -7,6 +7,7 @@ from typing import Optional
 from ycc_hull.db.entities import (
     BoatEntity,
     HolidayEntity,
+    LicenceInfoEntity,
     MemberEntity,
     MembershipTypeEntity,
     UserEntity,
@@ -60,11 +61,27 @@ class HolidayDto(CamelisedBaseModel):
         )
 
 
+class LicenceInfoDto(CamelisedBaseModel):
+    """
+    DTO for a YCC licence info.
+    """
+
+    id: int
+    licence: str
+
+    @staticmethod
+    def create(
+        licence_info: LicenceInfoEntity,
+    ) -> "LicenceInfoDto":
+        return LicenceInfoDto(id=licence_info.infoid, licence=licence_info.nlicence)
+
+
 class MemberPublicInfoDto(CamelisedBaseModel):
     """
     DTO for a member, containing information public to all active members.
     """
 
+    id: int
     username: str
     first_name: str
     last_name: str
@@ -78,6 +95,7 @@ class MemberPublicInfoDto(CamelisedBaseModel):
         member: MemberEntity,
     ) -> "MemberPublicInfoDto":
         return MemberPublicInfoDto(
+            id=member.id,
             username=member.user.logon_id,
             first_name=member.firstname,
             last_name=member.name,
@@ -100,17 +118,10 @@ class MemberSensitiveInfoDto(MemberPublicInfoDto):
     def create(
         member: MemberEntity,
     ) -> "MemberSensitiveInfoDto":
-        return MemberSensitiveInfoDto(
-            id=member.id,
-            username=member.user.logon_id,
-            first_name=member.firstname,
-            last_name=member.name,
-            membership_type=member.membership,
-            email=member.e_mail,
-            mobile_phone=member.cell_phone,
-            home_phone=member.home_phone,
-            work_phone=member.work_phone,
-        )
+        props = MemberPublicInfoDto.create(member).dict()
+        props["id"] = member.id
+        props["membership_type"] = member.membership
+        return MemberSensitiveInfoDto(**props)
 
 
 class MembershipTypeDto(CamelisedBaseModel):