+++ CHANGELOG.md
@@ -9,7 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add support for DB entities: licences
+- Add support for DB entities: licences, helpers
+- Add DTOs: licence info, helpers
+- Add test data: helpers
+- Add controller: helpers
+- Add API endpoints:
+  - GET helper task categories
+  - GET helper tasks
+  - GET helper task by ID
+  - POST subscribe as captain to a helper task
+  - POST subscribe as helper to a helper task
 
 ## [0.1.0] - 2023-04-05